// apps/pc/src/generated/models/Pl_bulletinpostsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_bulletinpostsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_bulletinpostsstatecode = keyof typeof Pl_bulletinpostsstatecode;
export const Pl_bulletinpostsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_bulletinpostsstatuscode = keyof typeof Pl_bulletinpostsstatuscode;

export interface Pl_bulletinpostsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_body?: string;
  pl_bulletinpostid: string;
  pl_contactkey?: string;
  pl_name?: string;
  "pl_PartnerLookup@odata.bind"?: string;
  pl_registeredat?: string;
  pl_requestkey?: string;
  statecode: Pl_bulletinpostsstatecode;
  statuscode?: Pl_bulletinpostsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_bulletinposts extends Pl_bulletinpostsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_partnerlookupname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_partnerlookup?: object;
  _pl_partnerlookup_value?: string;
}